+++ packages/babel-plugins/@glimmer/babel-plugin-strict-template-precompile/test/fixtures/custom-imports/output.js
@@ -1,6 +1,6 @@
 import Component from './component';
 ({
-  "id": "zK8QcLjf",
+  "id": null,
   "block": "[[[8,[32,0],null,null,null]],[],false,[]]",
   "moduleName": "(unknown template module)",
   "scope": () => [Component],