+++ packages/babel-plugins/@glimmer/babel-plugin-strict-template-precompile/test/fixtures/second-parameter/output.js
@@ -1,13 +1,13 @@
 import Component from './component';
 ({
-  "id": "zK8QcLjf",
+  "id": null,
   "block": "[[[8,[32,0],null,null,null]],[],false,[]]",
   "moduleName": "(unknown template module)",
   "scope": () => [Component],
   "isStrictMode": true
 });
 ({
-  "id": "zK8QcLjf",
+  "id": null,
   "block": "[[[8,[32,0],null,null,null]],[],false,[]]",
   "moduleName": "(unknown template module)",
   "scope": () => [Component],